refactor(components): header and nav bar
